+++ samples/Playground.Forms/Playground.Forms/Playground.Forms.csproj
@@ -13,6 +13,7 @@
     <ProjectReference Include="..\..\..\Softeq.XToolkit.Permissions\Softeq.XToolkit.Permissions.csproj" />
     <ProjectReference Include="..\..\..\Softeq.XToolkit.Remote\Softeq.XToolkit.Remote.csproj" />
     <ProjectReference Include="..\..\..\Softeq.XToolkit.WhiteLabel.Forms\Softeq.XToolkit.WhiteLabel.Forms.csproj" />
+    <ProjectReference Include="..\..\Playground.RemoteData\Playground.RemoteData.csproj" />
   </ItemGroup>
 
   <ItemGroup>
